Centralise widget information
Previously, we had a WidgetsProvider which provides the different types
of widgets for the site configuration. Additionally, the
WidgetRegistration class holds information about the widget registration
classes. This information is now centralised in one place: The Widgets
enum is now the place to modify the configuration when a new widget is defined.