Introduce RegistrationInterface for widget registration classes
Using the `AbstractRegistration` class as type hint in `Widgets` enum
reveals implementation details. Instead, an interface should be
used as only the `register()` method is necessary as API.

The `WidgetsTest` has enhanced to check every class if it implements
the `RegistrationInterface` - this is a "hard" check in contrast to the type
hint for the `getAssociatedClassNameForRegistration()` method.